💬 Update README
This commit updates the README to keep it accurate, but I guess you know
that? I'm removing the list of what I'd like to work on as I'm planning
to re-think what I want to learn again soon. Not now, though, but soon!

I'm basically having conversations with myself here, aren't I?